feat: add initialization data.sql file

+++ services/backend/warehouse-service/src/main/java/com/veteroch4k/warehouse/repositories/MaterialAccountingRepository.java
@@ -3,6 +3,6 @@ package com.veteroch4k.warehouse.repositories;
 import com.veteroch4k.warehouse.models.MaterialAccounting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialAccountingRepository extends JpaRepository<MaterialAccounting, Long> {
+public interface MaterialAccountingRepository extends JpaRepository<MaterialAccounting, Integer> {
 
 }
