+++ services/backend/warehouse-service/src/main/java/com/veteroch4k/warehouse/repositories/MaterialRepository.java
@@ -3,6 +3,6 @@ package com.veteroch4k.warehouse.repositories;
 import com.veteroch4k.warehouse.models.Material;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialRepository extends JpaRepository<Material, Long> {
+public interface MaterialRepository extends JpaRepository<Material, Integer> {
 
 }
